Add EEPROM selection for revision 5 boards.
The revision 5 boards use GPIO25 and GPIO28 to
enable the 2048 and 128 byte EEPROMs respectively.
This change adds a new command line option: -k (as
in 'khip select', because 'k' sounds like 'c'
sometimes, but not here... bleh. Running out of
letters.) to select between the three usable
states of these GPIOs:

  jcp -k 0: Disable both EEPROMs.
  jcp -k 1: Enable 128-byte/disable 2048-byte
  jcp -k 2: Enable 2048-byte/disable 128-byte

Disabling both is useful to lock out access if you
want to test some code or a ROM without corrupting
your EEPROM content. Enabling both at the same
time wouldn't make sense because they share the
same data lines, so they'd just fight for the bus
and corrupt each other.

The settings persist until the skunkboard is reset
(jcp -r or power-cycling the jaguar), so you can
select the EEPROM then proceed with flashing a ROM
or uploading code to memory that will use it:

  # Select the 2048-byte EEPROM
  jcp -k 2
  # Flash a game to bank 2 and start it
  jcp -2f biopede.rom

...

  # Select the 128-byte EEPROM
  jcp -k 1
  # Upload some code to memory and run it
  jcp my_homebrew.cof