Fix usb-drv-installer.exe Windows XP support
I patched the Windows XP support into libwdi,
avoiding the Zadig stuff that was more involved
and not used in the wdi-simple example project
that builds the little embedded driver installer
used here. The new usb-drv-installer.exe is built
with these patches included and has been
successfully tested on Windows XP (32-bit).

Also, add some notes on testing the installer and
some utilities to sign the driver installer and
the full installer package. Windows 10 is flagging
usb-drv-installer.exe as malware, and I verified
signing it fixes this (Used a self-signed CA for
testing purposes), so I'll need to get a real
certificate and sign it for the release version.