Add an Xcode project to build jcp on MacOS
To build, download libusb-0.1.12 and extract it
next to the top-level skunk_jcp directory. The
Xcode project's build will find it and take care
of configuring it and building it. The output
will be placed in ./xcode/{Release,Debug}
depending on the type of build run.

The build will create "universal binaries" for
both libusb and jcp itself, meaning they will run
natively on both Intel and ARM (Apple Silicon)
Macs.